# openapi_temp.json
{"openapi":"3.1.0","info":{"title":"SynapScale Backend","description":"API para gerenciamento de uploads de arquivos","version":"1.0.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health Check","description":"Endpoint para verificação de saúde da aplicação.\n\nReturns:\n    Status da aplicação","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["root"],"summary":"Root","description":"Endpoint raiz da aplicação.\n\nReturns:\n    Informações básicas da API","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/upload":{"post":{"summary":"Upload File","description":"Upload de arquivo.","operationId":"upload_file_api_v1_upload_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Limit"}},{"name":"window","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Window"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_api_v1_upload_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileUploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/":{"get":{"summary":"List Files","description":"Listar arquivos do usuário.","operationId":"list_files_api_v1__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Size"}},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/{file_id}":{"get":{"summary":"Get File","description":"Obter informações de um arquivo.","operationId":"get_file_api_v1__file_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete File","description":"Deletar arquivo.","operationId":"delete_file_api_v1__file_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/{file_id}/download":{"get":{"summary":"Download File","description":"Download de arquivo.","operationId":"download_file_api_v1__file_id__download_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileDownloadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/generate":{"post":{"tags":["llm"],"summary":"Generate Text","description":"Gera texto a partir de um prompt usando o provedor padrão ou especificado.\n\n- **prompt**: Texto de entrada para o modelo\n- **provider**: Provedor de LLM a ser usado (opcional)\n- **model**: Modelo específico a ser usado (opcional)\n- **max_tokens**: Número máximo de tokens a gerar (padrão: 1000)\n- **temperature**: Temperatura para amostragem (0.0-1.0) (padrão: 0.7)\n- **top_p**: Valor de top-p para amostragem nucleus (padrão: 0.95)\n- **top_k**: Valor de top-k para amostragem (padrão: 40)\n- **use_cache**: Se deve usar o cache (padrão: True)\n\nRetorna o texto gerado e metadados sobre a geração.","operationId":"generate_text_api_v1_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateTextRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Generate Text Api V1 Generate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/count-tokens":{"post":{"tags":["llm"],"summary":"Count Tokens","description":"Conta o número de tokens em um texto.\n\n- **text**: Texto para contar tokens\n- **provider**: Provedor de LLM a ser usado (opcional)\n- **model**: Modelo específico a ser usado (opcional)\n\nRetorna a contagem de tokens e metadados.","operationId":"count_tokens_api_v1_count_tokens_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CountTokensRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Count Tokens Api V1 Count Tokens Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/models":{"get":{"tags":["llm"],"summary":"List Models","description":"Lista todos os modelos disponíveis.\n\n- **provider**: Filtrar por provedor específico (opcional)\n\nRetorna a lista de modelos agrupados por provedor.","operationId":"list_models_api_v1_models_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"provider","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filtrar por provedor específico","title":"Provider"},"description":"Filtrar por provedor específico"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response List Models Api V1 Models Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/providers":{"get":{"tags":["llm"],"summary":"List Providers","description":"Lista todos os provedores disponíveis.\n\nRetorna a lista de provedores disponíveis e suas capacidades.","operationId":"list_providers_api_v1_providers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response List Providers Api V1 Providers Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/{provider}/generate":{"post":{"tags":["llm"],"summary":"Generate Text With Provider","description":"Gera texto a partir de um prompt usando o provedor especificado.\n\n- **provider**: Nome do provedor (ex: \"claude\", \"gemini\")\n- **prompt**: Texto de entrada para o modelo\n- **model**: Modelo específico a ser usado (opcional)\n- **max_tokens**: Número máximo de tokens a gerar (padrão: 1000)\n- **temperature**: Temperatura para amostragem (0.0-1.0) (padrão: 0.7)\n- **top_p**: Valor de top-p para amostragem nucleus (padrão: 0.95)\n- **top_k**: Valor de top-k para amostragem (padrão: 40)\n- **use_cache**: Se deve usar o cache (padrão: True)\n\nRetorna o texto gerado e metadados sobre a geração.","operationId":"generate_text_with_provider_api_v1__provider__generate_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"provider","in":"path","required":true,"schema":{"type":"string","description":"Nome do provedor","title":"Provider"},"description":"Nome do provedor"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateTextRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Generate Text With Provider Api V1  Provider  Generate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/{provider}/count-tokens":{"post":{"tags":["llm"],"summary":"Count Tokens With Provider","description":"Conta o número de tokens em um texto usando o provedor especificado.\n\n- **provider**: Nome do provedor (ex: \"claude\", \"gemini\")\n- **text**: Texto para contar tokens\n- **model**: Modelo específico a ser usado (opcional)\n\nRetorna a contagem de tokens e metadados.","operationId":"count_tokens_with_provider_api_v1__provider__count_tokens_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"provider","in":"path","required":true,"schema":{"type":"string","description":"Nome do provedor","title":"Provider"},"description":"Nome do provedor"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CountTokensRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Count Tokens With Provider Api V1  Provider  Count Tokens Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/{provider}/models":{"get":{"tags":["llm"],"summary":"List Models For Provider","description":"Lista os modelos disponíveis para o provedor especificado.\n\n- **provider**: Nome do provedor (ex: \"claude\", \"gemini\")\n\nRetorna a lista de modelos para o provedor.","operationId":"list_models_for_provider_api_v1__provider__models_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"provider","in":"path","required":true,"schema":{"type":"string","description":"Nome do provedor","title":"Provider"},"description":"Nome do provedor"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response List Models For Provider Api V1  Provider  Models Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_file_api_v1_upload_post":{"properties":{"category":{"type":"string","title":"Category"},"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["category","file"],"title":"Body_upload_file_api_v1_upload_post"},"CountTokensRequest":{"properties":{"text":{"type":"string","title":"Text","description":"Texto para contar tokens"},"provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Provider","description":"Provedor de LLM a ser usado"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model","description":"Modelo específico a ser usado"}},"type":"object","required":["text"],"title":"CountTokensRequest","description":"Modelo para requisição de contagem de tokens."},"FileDownloadResponse":{"properties":{"download_url":{"type":"string","title":"Download Url","description":"URL para download do arquivo"},"expires_at":{"type":"string","format":"date-time","title":"Expires At","description":"Data de expiração da URL"}},"type":"object","required":["download_url","expires_at"],"title":"FileDownloadResponse","description":"Esquema para resposta de download de arquivo.","example":{"download_url":"https://api.synapscale.com/download/123e4567-e89b-12d3-a456-426614174000?token=abc123","expires_at":"2023-01-01T13:00:00"}},"FileListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/FileResponse"},"type":"array","title":"Items","description":"Lista de arquivos"},"total":{"type":"integer","title":"Total","description":"Total de arquivos"},"page":{"type":"integer","title":"Page","description":"Página atual"},"size":{"type":"integer","title":"Size","description":"Tamanho da página"},"pages":{"type":"integer","title":"Pages","description":"Total de páginas"}},"type":"object","required":["items","total","page","size","pages"],"title":"FileListResponse","description":"Esquema para resposta de listagem de arquivos.","example":{"items":[{"category":"document","created_at":"2023-01-01T12:00:00","description":"Relatório financeiro anual","filename":"documento.pdf","id":"123e4567-e89b-12d3-a456-426614174000","is_public":false,"mime_type":"application/pdf","size":"1048576","status":"completed","tags":["relatório","financeiro"],"updated_at":"2023-01-01T12:00:00","user_id":"user_123"}],"page":1,"pages":1,"size":10,"total":1}},"FileResponse":{"properties":{"id":{"type":"integer","title":"Id"},"filename":{"type":"string","title":"Filename"},"original_filename":{"type":"string","title":"Original Filename"},"content_type":{"type":"string","title":"Content Type"},"file_size":{"type":"integer","title":"File Size"},"file_hash":{"type":"string","title":"File Hash"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags","default":[]},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"is_public":{"type":"boolean","title":"Is Public"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"}},"type":"object","required":["id","filename","original_filename","content_type","file_size","file_hash","is_public","created_at","updated_at"],"title":"FileResponse","description":"Esquema para resposta de arquivo.","example":{"category":"document","created_at":"2023-01-01T12:00:00","description":"Relatório financeiro anual","filename":"documento.pdf","id":"123e4567-e89b-12d3-a456-426614174000","is_public":false,"mime_type":"application/pdf","size":"1048576","status":"completed","tags":["relatório","financeiro"],"updated_at":"2023-01-01T12:00:00","user_id":"user_123"}},"FileUploadResponse":{"properties":{"file_id":{"type":"string","format":"uuid","title":"File Id","description":"ID único do arquivo"},"message":{"type":"string","title":"Message","description":"Mensagem de sucesso"}},"type":"object","required":["file_id","message"],"title":"FileUploadResponse","description":"Esquema para resposta de upload de arquivo.","example":{"file_id":"123e4567-e89b-12d3-a456-426614174000","message":"Arquivo enviado com sucesso"}},"GenerateTextRequest":{"properties":{"prompt":{"type":"string","title":"Prompt","description":"Texto de entrada para o modelo"},"provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Provider","description":"Provedor de LLM a ser usado"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model","description":"Modelo específico a ser usado"},"max_tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Tokens","description":"Número máximo de tokens a gerar","default":1000},"temperature":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Temperature","description":"Temperatura para amostragem (0.0-1.0)","default":0.7},"top_p":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Top P","description":"Valor de top-p para amostragem nucleus","default":0.95},"top_k":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Top K","description":"Valor de top-k para amostragem","default":40},"use_cache":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Use Cache","description":"Se deve usar o cache (se disponível)","default":true}},"type":"object","required":["prompt"],"title":"GenerateTextRequest","description":"Modelo para requisição de geração de texto."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"Bearer":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"Bearer":[]}]}